fix: Real fix for database validation issues
Apparently the issues mentioned in the previous commit were not due to issues in this mod; database changes in Civ 6 are cascaded, so when deleting an item all of its references are automatically deleted. However, what was happening was a DLC was then making changes to items deleted by this mod. To fix this, we have to set LoadOrder; I think the game and DLC have a LoadOrder of 0 and everything else gets loaded after, so as long as a number higher than 0 is set we should be okay. I used a much larger number to hopefully avoid compatibility issues with other mods as well.